CPANJMIS-master/CPANJMIS-master/project/requirements.txt
pillow
django
django-import-export==1.2.0
pycrypto
mysqlclient
PyMySQL
djangorestframework
django-guardian
pygments